requirements.txt
PyMuPDF==1.23.26
bs4==0.0.1
fastapi==0.110.1
langchain-community==0.0.9
langchain-core==0.1.7
langchain-openai==0.0.2
langchain==0.1.0
numexpr==2.8.7
openai==1.6.1
protobuf==5.26.1
python-dateutil==2.8.2
sentencepiece==0.2.0
tiktoken==0.5.2
torch==2.2.2
transformers==4.38.2
uvicorn==0.29.0
python-multipart==0.0.9
httpx==0.27.0
trl
datasets
hydra-core
omegaconf
peft
bitsandbytes